Refactor to single file
