+++ tests/test_ldiversity.py
@@ -1,4 +1,4 @@
-import ldiversity
+import anonymization
 import pandas as pd
 import pytest
 
@@ -8,10 +8,10 @@ import pytest
 
 def test_get_l_distinct():
     dataset = pd.read_csv("HW3.csv")
-    output = ldiversity.get_l_distinct(dataset, ['Sex'],'Drinks/Day')
+    output = anonymization.get_l_distinct(dataset, ['Sex'],'Drinks/Day')
     assert output == 6
 
 def test_get_l_entropy():
     dataset = pd.read_csv("HW3.csv")
-    output = ldiversity.get_l_entropy(dataset,['Sex'],'Drinks/Day')
+    output = anonymization.get_l_entropy(dataset,['Sex'],'Drinks/Day')
     assert round(output,5)==round(4.2266662750043125,5)
